@@ -30,9 +30,9 @@
     "@rollup/plugin-json": "4.1.0",
     "@rollup/plugin-node-resolve": "8.4.0",
     "ansi-regex": "5.0.0",
-    "colorette": "1.2.0",
+    "colorette": "1.2.1",
     "del": "5.1.0",
-    "eslint": "7.4.0",
+    "eslint": "7.5.0",
     "eslint-config-silverwind": "16.0.1",
     "execa": "4.0.3",
     "hosted-git-info": "3.0.5",
@@ -43,12 +43,12 @@
     "node-fetch": "2.6.0",
     "rc": "1.2.8",
     "registry-auth-token": "4.2.0",
-    "restana": "4.6.2",
-    "rollup": "2.21.0",
+    "restana": "4.7.0",
+    "rollup": "2.22.1",
     "rollup-plugin-hashbang": "2.2.2",
     "rollup-plugin-terser": "6.1.0",
     "semver": "7.3.2",
-    "tempy": "0.5.0",
+    "tempy": "0.6.0",
     "text-table": "0.2.0",
     "versions": "8.4.2"
   },
